+++ docs/release_notes.md
@@ -1,5 +1,9 @@
 ## DigiNode Tools Release Notes
 
+DigiNode Tools v0.9.3 - 2023-10-xx
+- Fix:
+
+
 DigiNode Tools v0.9.2 - 2023-10-23
 - Fix: Dashboard spacing when there is an update available in the Software section
 - Fix: Don't check the disk usage of IPFS if it is not present.